+++ ZenoDcimManager.Domain/ActiveContext/Commands/Inputs/CreateParameterCommand.cs
@@ -9,7 +9,7 @@ namespace ZenoDcimManager.Domain.ZenoContext.Commands.Inputs
         public string Unit { get; set; }
         public int LowLimit { get; set; }
         public int HighLimit { get; set; }
-        public int Scale { get; set; }
+        public double Scale { get; set; }
 
         public CreateParameterCommand()
         {
